Fix Typing Issues (#22)
Fixes a bunch of type issues identified by `mypy` and bumps minimum
python version to `3.10`. Also introduced `types.py` which can keep
track of types returned by `lotus` operations. There are still `mypy`
errors in the `models` directory but a lot of the model code will
undergo heavy rewrite anyways so leaving that for later. Once all the
`mypy` issues are resolved I will add it to the CI.